Refactor and pretty printing
I had originally envisioned this repo as being a temporary sandbox for
exploring a loose type system for Morloc. However, I think I may instead
keep it around and continually use it as a larger framework for
exploring a variety of type systems and type system implementations. As
I develop many systems under one umbrella, I can figure out clean ways
to factor out code that they share. Eventually, this repo can be a
stand-alone library for type systems and logic systems that are designed
to be hooked into from outside languages.

Basically, someone on the outside creates an AST and feeds it into this
type system. Then the algorithms here return the AST with richer
annotations.

TODO:

 * Add AsciiDoc documentation (rule syntax through LaTeX if possible)

 * Bidirectional.Simple:
 [ ] assign types to every subexpression of AST, not just the top one
 [ ] add color printing (red errors and green types)
 [ ] add type generator with QuickCheck
 [ ] add type generalization
 [ ] add let polymorphism
 [ ] add Int

 * Bidirectional.SystemF:
 [ ] add higher-kinded types (e.g. `List a` and `Map a b`)

 * Bidirectional.SystemF'
 [ ] SystemF plus typeclasses

 * More types
 [ ] tuples
 [ ] algebraic types
 [ ] GADTs
 [ ] records

 * Rule-based inference
 [ ] forward chaining algorithm

 * Morloc semantics
 [ ] casting
 [ ] properties (type inference within properties)
 [ ] language selection
 [ ] concrete case instantiation
 [ ] effect systems
 [ ] integration with ontologies and ruleset inference
 [ ] dependency handling