Basic monotypes with parameterization
Now I can type expressions such as:

```
(f :: [Num] -> [Bool]) [1,2,3]   |-   [Bool]
```

Next I need to generalize it to support polytypes, e.g.:

```
(f :: forall a . a -> a) 42   |-   Num
```

This will allow inference of map and other higher-order functions as
well, which, with regards to `morloc`, is the immediate goal of this
type checking program.

On second though, I think I'll put an end to this branch of xi and focus
on implementing Dunfield's algorithm.