Added exit log and found logic bug
The problem was a simple off-by-1 error that resulted in dropping the
first element of context in each cut. Now it is fixed.

There is a huge amount of output for a simple little type checking
operation like `(\x.x)Unit`, I should next comb through the output in
detail to ensure all is well. Then I can add in more primitives and
maybe let syntax, then I'll be able to test more complex expressions.