Saw the function contravariance light
My prior assumption that Dunfield had the subtype argument for functions
wrong was based on my incorrect conflation of the concepts of subtyping
and inheritance.

Function x can be used in the same context as function y (i.e. x <: y)
if the inputs to x are more inclusive than y and the outputs more
exclusive.

Suppose we have a expression `h(g(f x))`. We want to know what
functions can replace g. That is, we want to find `g'` where `g'<:g`. The
domain of `g'` must influde all possible outputs of `f` and the codomain
of `g'` (the output) must be within the domain of `h`. Writing out all
the type signatures:

```
x :: A
f :: A -> B
g :: B -> C
h :: C -> D
g :: B' -> C' where (x in B for x in B') and (x in C' for x in C)
```

So (B'->C' <: B->C) if (B<:B') and (C'<:C).