Fix InstRArr and add lambda argument sugar
The InstRArr rule uses the syntax `\Gamma[a2,a1,a=a1->a2]` which I had
been interpretting as `\Gamma,a2,a1,a=a1->a2` (as in `-->I=>`).
Actually, this syntax was intended to means:

  (G, a, G') --> (G, a2, a1, a=a1->a2, G')

That is, the whole chunk [a2,a1,a=a1->a2] is inserted into context.

I also add sugar for writing expressions of the form: `\x y -> x`;
multiple parameters passed to a lambda without explicitly nesting.