Add stub support for statement syntax
I can parse and run statements like "x=1; 4.2". The typechecker
currently ignores all be the last line (I'll add in that logic next). A
program then is comprised of a series of declarations followed by a
terminal expression. Each declaration is generalized and stored in
context. Redefining a varialbe is illegal.