Generalize and better verbosity handling
I now quantify over existentials as a final step. Thus

```
f = (\x y -> x); f 42
_ :: Forall a . a -> Int
```

Also, verbosity is now an integer, rather than a boolean, giving me
finer control. I need to flesh out the verbosity levels, though. Maybe

  0) nothing printed (system flag indicates success/failure)
  1) resulting type is printed
  2) the whole shebang

I sunk a few hours on a deadend attempt to make the declarations order
invariant. Statements like `x=1; y=f x` and `y=f x; x=1` should both be
legal and identical. However, implementing this in the ordered context
model proved counterintuitive and convoluted.

Taking `y=f x; x=1` as an example, simply stepping through the statements
in order leads to an unbound variable error in `f x`, where `x` is not
defiend. The algorithm cannot see the future context. I tried doing a
first pass which would initialize each declaration but not typecheck the
right hand expression. Dealing with usage of undefined variables though
means my inference functions cannot return complete types and statements
cannot be generalized immediately after they are typed.

I tried storing the right-hand expression along with the left-hand
variable name during the first pass. Then a function that uses the
variable could call `infer` on the left-hand expression. However, this
deeply convoluted the code. The `apply` function, where context is
applied to a type to replace unsolved existentials with existing
solutions, would have to be able to lookup declarations in addition to
solved existentials. If the declaration has not been typed (i.e. the
statement is to the right of the current one), then `infer` is called,
returning a new context and type. Thus the simple signature
`apply::Gamma->Type->Type` becomes
`apply::Gamma->Type->Stack(Gamma,Type)`. Applying a context gained the
side-effect of altering the context.

The infer-Var rule also needed to be able to lookup the variable to see
if it was defined, and if not to call `infer` on the associated
expression.

All of these `infer` calls must be independent of the contents of the
local calling scope (at least if I want lexical scope, which I do). That
means I need to call `infer` with a restricted context. For this
purpose, we would need to add a statement marker, then cut all context
up to the marker and pass the remainder to `infer e`. Then the resulting
context would have to be restricted and appended to the local context
(before the statement marker). Then there is the issue of mutal
recursion (and non-mutual, for that matter).

I think it might make more sense to implement the collection of
declarations as an unordered container linking names to expressions: an
unordered map of ordered lists. I'm not quite sure yet.

The constraint that every expression must be defined before it is used
allows programs (collections of statements) to be implemented using
nested `let` expressions. This is easy to implement and fits well with
the ordered-context approach.

I'll table declaration order invariance for the moment and start working
on parametric polymorphism.