Higher-kinded types are now online
Now I can write:

```
apply :: forall a b . (a->b) -> a -> b;
f :: Int -> Bool;
apply f 42;
```

The above should arbuably fail, since `apply` and `f` are unbound.
However, I added a case to `infer` for annotated variables (signatures)
that ignores this quibble. Thus, if a single variable is annotated, e.g.
`x::Int`, and is not declared, this would normally raise an
UnboundVariable error. However, it is convenient for testing purposes,
and also for Morloc where functions are imported as black boxes from
other langugaes, to be able to simply declare a type as an axiom.
Perhaps I should add dedicated syntax for axiomatic type declarations?