9/26 tests fail - return annotated expressions
This change really convoluted the code. And I find myself guessing about
when to apply generalization and context application to the expressions.
The beauty of the old implementation, where the code was only slightly
less transparent than the rules, is lost. I wonder if there is a way to
regain this by building the expression within the monad?

But first I want my tests to pass. Once they do, I can start testing the
hell out of the system and begin to gain a deeper intuition into what
results should look like.

I also need to consider invariants and start building quickcheck tests.