(2/26 fail) more bug fixes, feeling apprehensive
Some of the test that pass (i.e. have a correct type for the final
statement), have incorrectly annotated internal types. For example:

```
xi '((\x -> x) :: forall a . a -> a) True'
```

produces

```
((\x -> (x :: a) :: forall a . a -> a) (True :: forall a . a -> a) :: Bool)
```

specifically, `True` here is given the type of the function of which it
is a part. The same problem would appears when other primitive types are
switched for the Bool.