Add readType and generalize quickcheck test
I am considering removing IO from the monad stack. I don't really need
it since the typechecker is not interacting with the environment in any
way except printing. The only reason I print is to see what is happening
internally, rather than to get a real-time updates. So instead of
printing I could just store the log messages in the Writer monad and
print them after the run.