Remove IO from Stack and remove prettyprint
I will likely want to add the diagnostics and prettyprint infrastructure
back (in some form) in the future. But for now it is just clutter. They
were all useful when I was diagnosing problems interactively, but less
so in the property testing phase.

I removed the IO monad. But after removing diagnostics, I might as well
reduce the stack down to just an Either monad.