Update docs for list and tuple
I have been looking for an approach to integrate this into Morloc.
Currently Morloc and Xi both have parsers, but they are slowly
converging and should ultimately be merged. There should be one Morloc
parser that generates Xi expressions. Xi typechecks them returning fully
annotated expressions. Morloc converts these expressions to Manifolds,
which then are used to generate code.

Logically, Xi should also be the system that interacts with MorlocIO to
tap external modules and access community knowledge when selecting
functions. It will need a logic engine.

                     .--> kita <----------.
            upload  /   (library)          \ reuse binaries
                   /         \              \
                  /           \   compile    \
                nishi ---------+----------> higashi
               (logic)        /            (execution)
                             /                /
                          minami  <----------'
                       (knowledge)       runtime and usage stats

Organizing by direction is visually convenient. xi is Chinese for West,
but I really don't want to use the Chinese word for East (dong). So
Japanese is good. kita and minami are not implemented but will reside
within MorlocIO. Xi is nishi and the current Morloc codebase is mostly
higashi (along with connective tissue that should possibly serve as a
fifth position).

TODO:
    [ ] add more failing tests in QC
    [ ] reread DK paper and proof section - implement any invariants in QC
    [ ] read Simon Peyton Jones papers, extend docs and add refs
    [ ] add haddock docs at function and module level
    [ ] add language-specific `source`
    [ ] make unbound variables an error in all circumstances
    [ ] add modules, with import, export, and namespaces (with rules and tests)
    [ ] make terminal expressions optional (scripts have, modules haven't)
    [ ] sort within modules -- signatures < declarations < expression
    [ ] die upon re-declaration in all circumstances
    [ ] add language-specific type signatures with subtyping rules
    [ ] add properties - for now only used to label (un)pack and cast functions
    [ ] identify serialization/deserialization functions
    [ ] build call graph
    [ ] infer specific-language, when possible
    [ ] insert serialization steps as needed
    [ ] identify cast functions
    [ ] insert casts as needed to resolve type conflicts
    [ ] add constraint syntax (for compatibility, no semantics yet)