Add tests for NRank types (which aren't supported)
I do not intend to support NRank types, so the following expression
should fail:

```
f :: forall a . a -> a
g h = (h 42, h True)
g f
```

When `f` is passed to `g`, it collapses to a monotype. Thus cannot
support both the applications `h 42` and `h True`. Haskell has an
extensions that allows this (n-rank types), but I don't want to support
it.

Back to Xi, I wrote a new unit test that asserts failure of the above
expression. This was successful. But while writing a similar expression
that should pass, I found an infinite loop:

```
f :: forall a . a -> a
g = \h -> (h 42, h 56)
g f
```

Shrinking to `f :: forall a . a -> a; g = \h->h; g f` still creates an
infinite loop.

Also, this terminates:

```
f :: forall a . a -> a; g :: forall b . b -> Int; g f
```

While, this does not:

```
f :: forall a . a -> a; g :: forall a . a -> Int; g f
```

The difference only being the reuse of the bound variable `a`.  The
QuickCheck tests miss such cases since the type generator does not reuse
variable names.