Allow signatures and declarations
I've added tests for proper checking as well. For example, this passes:

```
f :: Int -> Bool
f x = True
f 42
```

Whereas this fails with a SubtypeError:

```
f :: Int -> Bool
f x = 9999
```

The Declaration and Signature constructors no longer take expressions
containing the rest of the program recursively. Now a Program is just a
list of Expr's. This means that it is no longer necessary to add a
terminal expression that is not a signature or declaration.

One slightly annoying thing though is that I sort the expressions,
putting signatures and declarations before other things. This doesn't
change the meaning of the code, but can be confusing when printing
results.

I had to kill my Expr generator. I will need to write a new one
eventually. I would like it to be capable of generating correctly typed
expressions, probably by mapping arbitrary types to arbitrary
expressions. The old implementation was hacky as hell.