Add import/export context restriction
The system seems to be working, though I think we could make better
error messages. Specifically, a error is raised if an imported function
does not exist. There are two possible reasons for this, 1) the module
might not export the variable or 2) the module might not define the
variable. Currently both cases raise the same error.