Store type extensions in AnnG
Where should the type extensions (concrete type instances, properties
and constraints) be stored?

For now I am storing them only at the top-level AnnG type constructor in
the context. Realization of a module will require selection of a
realization for each function that is called from the realizations
stored in context.

I will return to this problem eventually, but first I want to add
handling for merging realizations with the abstract type into a
"compound type" (or whatever we call it).

What constraints are their between the abstraction and realization? I
can only think of one that will always hold: they must have the same
number of arguments (the same kind). There may be properties and
constraints that are specific to the realization.

Without specific knowledge about the property, I can't say whether it
should be inherited. `pack`, `unpack`, and `cast` should all be
inherited.

Constraints should map in some way to the realizations, though for now I
am not parsing them.