Add scientific numbers remove integers
Now all numbers are Num, which is an arbitrary accuracy scientific
number with an Int exponent (e.g., 64 bit).

I had to remove all Int from the tests. This brings up an interesting
problem, since I really do like to use the type Int, even if everything
under the hood has the same number representation.

I think it is good for all numbers to be typed as "Num". But the
programmer should be able to cast these to Int, Integer, Nat or whatever
else. Most languages have this handling built in, which allows typesafe
addition of integers and doubles.

However, in Morloc I do not want to hardcode this. Rather I would prefer
the general, programmable, cast rules handle it.

These rules must be at least strong enough for me to implement the
Haskell-style of handling number conversions.

But casting is for later.

I also need type definitions, e.g.:

```
type Int = i:Num where { i - (floor i) == 0 }
type Nat = i:Int where { i >= 0 }
```

These are subset operations. From the types we can infer that Nat is a
subset of Int, thus any Nat can be cast as an Int. The reverse, casting
an Int as a Nat, requires asserting the conditions, which may or may not
pass.

But typedefs are also for later ...