Add line and block comments
I'm going with haskell comments: line comments of `--` and block
comments starting with `{-` and ending with `-}`. I slightly prefer the
Haskell syntax over the heavier `#`. I would also like to reserve `#`
for macros or pragmas. I like the Algol `//`, but am not fond of the
asymetric `/* */` block comments.

This is a deviation from the current Morloc syntax.