Allow code to be read from files (and bug fixes)
 * can read from files with `xi -f foo.loc`
 * allow semicolons at the end of a program
 * fix import bug (needed commaSep1 not many1)

I wrote an example program that defines and uses a map function. Looking
over the annotated output, I am reminded that I need to write a
`realize` function that replaces all general terms with specific ones.
Eventually, this task should be delegated to a wiser entity (the morloc
AI, who will model performance and check for dependencies). But for now,
I should write a simple rule-based function that minimizes the number of
language transitions.