Generalize API to take file handlers as args
The API now consists of the following 4 functions: `parse`, `typecheck`,
`cute` and `ugly`.

The `cute` and `ugly` functions are for pretty and raw printing of the
modules (mostly just for diagnostics at this point).

The `parse` and `typecheck` have the following signatures:

```
parse
    :: (FileName -> IO ())
    -> (MVar -> IO Text)
    -> Text
    -> IO [Module]
typecheck
    :: [Module]
    -> Either TypeError [Module]
```

The first two arguments to `parse` check for the existence of sourced
files and import the code for modules. These functions are far more than
just file readers, though. They are can be modified to, for example
 * lookup files in the working directory
 * search through customized system paths
 * search online databases

These functions are the interface between the pure type handling code
and the rest of Morloc. They are not quite complete, though, since I
will eventually need to be able to read module metadata and thread state
through the typechecker. The state could include things such as a list
of supported languages, strictness settings or verbosity settings.