Store source and import aliases without Maybe
If no alias for an imported or sourced term, then the given name is
used. So rather than return nothing when no alias is given, we can just
use the first name. No information is lost and the code is simplified.
The tuple returned contains the source name and the morloc name, which
may happen to be the same.